assert and cast
