+++ ddprof-lib/src/main/cpp/itimer.cpp
@@ -21,7 +21,7 @@
 #include "os.h"
 #include "profiler.h"
 #include "stackWalker.h"
-#include "thread.h"
+#include "threadLocalData.h"
 #include "threadState.inline.h"
 #include "guards.h"
 #include <sys/time.h>
