+++ ddprof-lib/src/main/cpp/objectSampler.cpp
@@ -12,7 +12,7 @@
 #include "objectSampler.h"
 #include "pidController.h"
 #include "profiler.h"
-#include "thread.h"
+#include "threadLocalData.h"
 #include <jni.h>
 #include <limits.h>
 #include <math.h>
